added employee service and rest controller

+++ src/main/java/com/hdfc/olms/dto/EmployeeDTO.java
@@ -9,7 +9,7 @@ import lombok.NoArgsConstructor;
 @Data
 public class EmployeeDTO {
 	
-	private int employeeId;
+	private long employeeId;
 	private String name;
 	private String email;
 	private String phoneNumber;
